# swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Beautiful Occupation API 
consumes:
  - application/json
produces:
  - application/json
basePath: /api/v1/
...
paths:
...
  /current_user:
    post:
      operationId: controller.auth.get_user
      tags: 
        - Auth
      summary: Get current user info from json web token
      parameters:
        - name : access_token
          in : body
          required : True
          schema :
            type : object
            properties:
              jwt : 
                type: string
      responses:
        200:
          description : Successfully returned the json web token
          schema :
            type : object
            properties:
                fname:
                  type: string
                lname:
                  type: string
                occupation:
                  type: string
                location:
                  type: string
                cell:
                  type: string
                timestamp:
                  type: string
...